rename testreport handler file
