+++ lib/testreport_handler.rb
